deploy-model/requirements.txt
streamlit
pandas
joblib
numpy
tensorflow-cpu